chores: 新增 server.js, 將 socket 從 app.js 中移出
