+++ .env-template
@@ -2,8 +2,8 @@ API_VERSION = '1.0'
 TAPPAY_PARTNER_KEY = ''
 TAPPAY_MERCHANT_ID = 'AppWorksSchool_CTBC'
 NODE_ENV = 'development'
-PORT = 3000
-PORT_TEST = 3001
+PORT = 8000
+PORT_TEST = 8001
 TOKEN_EXPIRE = 2592000
 TOKEN_SECRET = ''
 BCRYPT_SALT = 10
